refactor: improve SettingsPanel structure and maintainability
- Split SettingsPanel into modular components
- Extract common CSS styles
- Centralize configuration options and constants
- Implement useConfirmation composable for confirmation dialog logic
- Reduce code duplication and improve separation of concerns
