+++ src/assets/theme.css
@@ -132,3 +132,89 @@
   margin-top: 1rem;
   animation: blink 1s step-end infinite;
 }
+
+/* Settings panel styles */
+.settings-section {
+  display: flex;
+  flex-direction: column;
+  gap: 0.8rem;
+  padding: 1rem;
+  border: 2px solid var(--button-border);
+  background-color: rgba(20, 20, 20, 0.8);
+  position: relative;
+  overflow: hidden;
+}
+
+.settings-section::before {
+  content: '';
+  position: absolute;
+  top: 0;
+  left: 0;
+  width: 100%;
+  height: 100%;
+  background: linear-gradient(45deg, transparent 0%, rgba(51, 255, 51, 0.05) 50%, transparent 100%);
+  z-index: 0;
+}
+
+.section-title {
+  font-family: var(--font-display);
+  font-size: 1rem;
+  margin: 0;
+  color: var(--text-primary);
+  border-bottom: 1px solid var(--text-primary);
+  padding-bottom: 0.5rem;
+  margin-bottom: 0.5rem;
+  z-index: 1;
+}
+
+.setting-label {
+  font-family: var(--font-mono);
+  font-size: 1.1rem;
+  color: var(--text-secondary);
+}
+
+.setting-item {
+  display: flex;
+  flex-direction: column;
+  gap: 0.5rem;
+  z-index: 1;
+  margin-bottom: 0.8rem;
+}
+
+.setting-control {
+  display: flex;
+  gap: 0.5rem;
+}
+
+/* Option button styles */
+.option-button {
+  padding: 0.4rem 0.6rem;
+  background-color: var(--button-bg);
+  color: var(--text-secondary);
+  font-family: var(--font-mono);
+  font-size: 0.9rem;
+  border: 1px solid var(--button-border);
+  cursor: pointer;
+  flex: 1;
+  text-align: center;
+}
+
+.option-button.selected {
+  background-color: var(--primary);
+  color: #000;
+  border-color: var(--primary);
+}
+
+/* Status colors */
+.warning {
+  color: var(--warning);
+}
+
+.danger {
+  color: var(--danger);
+}
+
+.action-text {
+  font-family: var(--font-mono);
+  font-size: 1.1rem;
+}
